+++ socid_extractor/main.py
@@ -9,8 +9,10 @@ from .postprocessor import POSTPROCESSORS
 
 HEADERS = {
     'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36',
-    "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9"
+    "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
 }
+
+PROCESS_ERRORS = (AttributeError, KeyError, IndexError, TypeError)
 
 
 def parse_cookies(cookies_str):
@@ -50,6 +52,31 @@ def mutate_url(url):
     return mutate_results
 
 
+def transform(scheme_data, extracted_data):
+    transforms = scheme_data.get('transforms', [])
+    if transforms:
+        for t in transforms:
+            logging.debug(t)
+            try:
+                extracted_data = t(extracted_data)
+            except PROCESS_ERRORS as e:
+                logging.debug(f'Transform error: {e}')
+                extracted_data = {}
+            logging.debug(extracted_data)
+    return extracted_data
+
+
+def map_fields(scheme_data, transformed_data):
+    values = {}
+    for name, get_field in scheme_data['fields'].items():
+        try:
+            value = get_field(transformed_data)
+            values[name] = str(value) if value not in (None, [], {}) else ''
+        except PROCESS_ERRORS as e:
+            logging.debug(f'Unable to extact field {name}: {e}')
+    return values
+
+
 def extract(page):
     for scheme_name, scheme_data in schemes.items():
         flags = scheme_data['flags']
@@ -78,21 +105,11 @@ def extract(page):
 
             if scheme_data.get('extract_json', False):
                 extracted = regexp_group.group(1)
-
                 logging.debug('Extracted: %s', extracted)
 
-                transforms = scheme_data.get('transforms', [])
-                if transforms:
-                    for t in transforms:
-                        logging.debug(t)
-                        try:
-                            extracted = t(extracted)
-                        except (AttributeError, KeyError) as e:
-                            logging.debug(f'Transform error: {e}')
-                            extracted = {}
-                        logging.debug(extracted)
+                transformed = transform(scheme_data, extracted)
 
-                json_data = json.loads(extracted)
+                json_data = json.loads(transformed)
 
                 if json_data == {}:
                     logging.debug('Unabled to extract json!')
@@ -105,14 +122,17 @@ def extract(page):
                     with open('debug_extracted.json', 'w') as f:
                         f.write(loaded_json_str)
 
-                for name, get_field in scheme_data['fields'].items():
-                    try:
-                        value = get_field(json_data)
-                        values[name] = str(value) if value not in (None, [], {}) else ''
-                    except (AttributeError, KeyError, IndexError) as e:
-                        logging.debug(f'Unable to extact field {name}: {e}')
+                values = map_fields(scheme_data, json_data)
             else:
-                values = regexp_group.groupdict()
+                groupdict = regexp_group.groupdict()
+                if groupdict:
+                    values = groupdict
+                else:
+                    extracted = regexp_group.group(1)
+                    logging.debug('Extracted: %s', extracted)
+
+                    transformed_data = transform(scheme_data, extracted)
+                    values = map_fields(scheme_data, transformed_data)
 
         if use_html_parser:
             soup = bs(page, 'html.parser')
@@ -120,14 +140,14 @@ def extract(page):
                 try:
                     value = get_field(soup)
                     values[name] = str(value) if value != None else ''
-                except (AttributeError, KeyError, IndexError) as e:
+                except PROCESS_ERRORS as e:
                     logging.debug(f'BS extract error: {e}')
 
         for p in POSTPROCESSORS:
             try:
                 additonal_data = p(values).process()
                 values.update(additonal_data)
-            except Exception as e:
+            except PROCESS_ERRORS as e:
                 logging.debug('Postprocess error: ', e)
 
         return {a: b for a, b in values.items() if b or type(b) == bool}